requirements.txt
python-dotenv
google-genai
google-generativeai 
firebase_admin
json
backend
flask
pydantic
